==> 1_read_newdata.ipynb <==
import pandas as pd
import glob
import dateparser
import re
from striprtf.striprtf import rtf_to_text

def parser(file):
 
    # Open a TXT file. Store all articles in a list. Each article is an item
    # of the list. Split articles based on the location of such string as
    # 'Document PRN0000020080617e46h00461'
 
    articles = []
    i = 1
    filename = []
    with open(file, 'r') as infile:
        data = infile.read()
    start = re.search(r' \n', data).start()
    for m in re.finditer(r'Dokument [a-zA-Z0-9]{25}\n', data):
        end = m.end()
        a = data[start:end].strip()
        a = '\n   ' + a
        name = file + f"_{i}"
        articles.append(a)
        filename.append(name)
        start = end
        i+=1
    
    return articles, filename

def findbetween(text, st, en):

    start = re.search(st, text).end()
    end = re.search(en, text).start()
    a = text[start:end].strip()
    
    return a

def findtime(text, st, en):

    start = re.search(st, text).start()
    end = re.search(en, text).end()
    a = text[start:end].strip()
    
    return a

### For Factiva

news = []
time = []
names = []

for files in glob.glob("newsarticle/Sunday Times/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Sunday Times\n")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

Sunday_Times = pd.DataFrame()
Sunday_Times["filename"] = names
Sunday_Times["text_original"] = news
Sunday_Times["date"] = time
Sunday_Times["year"] = Sunday_Times["date"].dt.year
Sunday_Times["month"] = Sunday_Times["date"].dt.month
Sunday_Times["outlet"] = "Sunday Times"
Sunday_Times["country"] = "South Africa"
Sunday_Times["text_clean"] = clean_text
Sunday_Times["South"] = 1

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Globe & Mail/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Globe and Mail\n")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

Globe_Mail = pd.DataFrame()
Globe_Mail["filename"] = names
Globe_Mail["text_original"] = news
Globe_Mail["date"] = time
Globe_Mail["year"] = Globe_Mail["date"].dt.year
Globe_Mail["month"] = Globe_Mail["date"].dt.month
Globe_Mail["outlet"] = "GlobeMail"
Globe_Mail["country"] = "Canada"
Globe_Mail["text_clean"] = clean_text
Globe_Mail["South"] = 0

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Hindu/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Hindu\n")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

Hindu = pd.DataFrame()
Hindu["filename"] = names
Hindu["text_original"] = news
Hindu["date"] = time
Hindu["year"] = Hindu["date"].dt.year
Hindu["month"] = Hindu["date"].dt.month
Hindu["outlet"] = "Hindu"
Hindu["country"] = "India"
Hindu["text_clean"] = clean_text
Hindu["South"] = 1

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Nation/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Nation.*\nTHENAT")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

The_Nation = pd.DataFrame()
The_Nation["filename"] = names
The_Nation["text_original"] = news
The_Nation["date"] = time
The_Nation["year"] = The_Nation["date"].dt.year
The_Nation["month"] = The_Nation["date"].dt.month
The_Nation["outlet"] = "The Nation"
The_Nation["country"] = "Thailand"
The_Nation["text_clean"] = clean_text
The_Nation["South"] = 1

news = []
time = []
names = []

for files in glob.glob("newsarticle/The New Zealand Herald/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\n.*New Zealand Herald\n")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

NZ_Herald = pd.DataFrame()
NZ_Herald["filename"] = names
NZ_Herald["text_original"] = news
NZ_Herald["date"] = time
NZ_Herald["year"] = NZ_Herald["date"].dt.year
NZ_Herald["month"] = NZ_Herald["date"].dt.month
NZ_Herald["outlet"] = "The New Zealand Herald"
NZ_Herald["country"] = "New Zealand"
NZ_Herald["text_clean"] = clean_text
NZ_Herald["South"] = 0

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Star/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Star\nTHESTR")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

Star = pd.DataFrame()
Star["filename"] = names
Star["text_original"] = news
Star["date"] = time
Star["year"] = Star["date"].dt.year
Star["month"] = Star["date"].dt.month
Star["outlet"] = "The Star"
Star["country"] = "South Africa"
Star["text_clean"] = clean_text
Star["South"] = 1

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Times/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Times\nT")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

Times = pd.DataFrame()
Times["filename"] = names
Times["text_original"] = news
Times["date"] = time
Times["year"] = Times["date"].dt.year
Times["month"] = Times["date"].dt.month
Times["outlet"] = "The Times"
Times["country"] = "UK"
Times["text_clean"] = clean_text
Times["South"] = 0

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Washington Post/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Washington Post\nWP")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

WPost = pd.DataFrame()
WPost["filename"] = names
WPost["text_original"] = news
WPost["date"] = time
WPost["year"] = WPost["date"].dt.year
WPost["month"] = WPost["date"].dt.month
WPost["outlet"] = "The Washington Post"
WPost["country"] = "USA"
WPost["text_clean"] = clean_text
WPost["South"] = 0

news = []
time = []
names = []

for files in glob.glob("newsarticle/Times of India/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Times of India\nTOI")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

Times_IN = pd.DataFrame()
Times_IN["filename"] = names
Times_IN["text_original"] = news
Times_IN["date"] = time
Times_IN["year"] = Times_IN["date"].dt.year
Times_IN["month"] = Times_IN["date"].dt.month
Times_IN["outlet"] = "Times of India"
Times_IN["country"] = "India"
Times_IN["text_clean"] = clean_text
Times_IN["South"] = 1

news = []
time = []
names = []

for files in glob.glob("newsarticle/Toronto Star/**/*.txt", recursive = True):  
    articles, name = parser(files)
    news.extend(articles)
    names.extend(name)
    
for n in news:
    results = findbetween(n, "Wöter\n", "\nThe Toronto Star\nTOR")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 

TO_Star = pd.DataFrame()
TO_Star["filename"] = names
TO_Star["text_original"] = news
TO_Star["date"] = time
TO_Star["year"] = TO_Star["date"].dt.year
TO_Star["month"] = TO_Star["date"].dt.month
TO_Star["outlet"] = "Toronto Star"
TO_Star["country"] = "Canada"
TO_Star["text_clean"] = clean_text
TO_Star["South"] = 0

### For Nexis

news = []
time = []
names = []

for files in glob.glob("newsarticle/Bangkok Post/**/*.rtf", recursive = True):
    with open(files, "r") as fi:
        rtf = fi.read()    
        text = rtf_to_text(rtf)
        news.append(text)
        names.append(files)
    
for n in news:
    results = findbetween(n, "\nThe Bangkok Post.*\n", "\n\n\nCopyright")
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 
    
Bangkok_PO = pd.DataFrame()
Bangkok_PO["filename"] = names
Bangkok_PO["text_original"] = news
Bangkok_PO["date"] = time
Bangkok_PO["year"] = Bangkok_PO["date"].dt.year
Bangkok_PO["month"] = Bangkok_PO["date"].dt.month
Bangkok_PO["outlet"] = "Bangkok Post"
Bangkok_PO["country"] = "Thailand"
Bangkok_PO["text_clean"] = clean_text
Bangkok_PO["South"] = 1

news = []
time = []
names = []

for files in glob.glob("newsarticle/Sydney Morning Herald/**/*.rtf", recursive = True):
    with open(files, "r") as fi:
        rtf = fi.read()    
        text = rtf_to_text(rtf)
        news.append(text)
        names.append(files)
    
for n in news:
    results = findtime(n, "\n\w* \d{1,2}, \d{4}", "\d{4} \w*day\n")
    
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 
    
Sydney_MO = pd.DataFrame()
Sydney_MO["filename"] = names
Sydney_MO["text_original"] = news
Sydney_MO["date"] = time
Sydney_MO["year"] = Sydney_MO["date"].dt.year
Sydney_MO["month"] = Sydney_MO["date"].dt.month
Sydney_MO["outlet"] = "The Sydney Morning Herald"
Sydney_MO["country"] = "Australia"
Sydney_MO["text_clean"] = clean_text
Sydney_MO["South"] = 0

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Australian/**/*.rtf", recursive = True):
    with open(files, "r") as fi:
        rtf = fi.read()    
        text = rtf_to_text(rtf)
        news.append(text)
        names.append(files)
    
for n in news:
    results = findtime(n, "\n\w* \d{1,2}, \d{4}", "\d{4} \w*day\n")
    
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 
    
The_Australian = pd.DataFrame()
The_Australian["filename"] = names
The_Australian["text_original"] = news
The_Australian["date"] = time
The_Australian["year"] = The_Australian["date"].dt.year
The_Australian["month"] = The_Australian["date"].dt.month
The_Australian["outlet"] = "The Australian"
The_Australian["country"] = "Australia"
The_Australian["text_clean"] = clean_text
The_Australian["South"] = 0

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Guardian/**/*.rtf", recursive = True):
    with open(files, "r") as fi:
        rtf = fi.read()    
        text = rtf_to_text(rtf)
        news.append(text)
        names.append(files)
    
for n in news:
    results = findtime(n, "\n\w* \d{1,2}, \d{4}", "\d{1,2}, \d{4}")
    
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 
    
The_Guardian = pd.DataFrame()
The_Guardian["filename"] = names
The_Guardian["text_original"] = news
The_Guardian["date"] = time
The_Guardian["year"] = The_Guardian["date"].dt.year
The_Guardian["month"] = The_Guardian["date"].dt.month
The_Guardian["outlet"] = "Guardian"
The_Guardian["country"] = "UK"
The_Guardian["text_clean"] = clean_text
The_Guardian["South"] = 0

news = []
time = []
names = []

for files in glob.glob("newsarticle/The New York Times/**/*.rtf", recursive = True):
    with open(files, "r") as fi:
        rtf = fi.read()    
        text = rtf_to_text(rtf)
        news.append(text)
        names.append(files)
    
for n in news:
    results = findtime(n, "\n\w* \d{1,2}, \d{4}", "\n\w* \d{1,2}, \d{4}")
    
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 
    
NY_Times = pd.DataFrame()
NY_Times["filename"] = names
NY_Times["text_original"] = news
NY_Times["date"] = time
NY_Times["year"] = NY_Times["date"].dt.year
NY_Times["month"] = NY_Times["date"].dt.month
NY_Times["outlet"] = "The New York Times"
NY_Times["country"] = "USA"
NY_Times["text_clean"] = clean_text
NY_Times["South"] = 0

news = []
time = []
names = []

for files in glob.glob("newsarticle/The Press/**/*.rtf", recursive = True):
    with open(files, "r") as fi:
        rtf = fi.read()    
        text = rtf_to_text(rtf)
        news.append(text)
        names.append(files)
    
for n in news:
    results = findtime(n, "\n\w* \d{1,2}, \d{4}", "\n\w* \d{1,2}, \d{4}")
    
    date = dateparser.parse(results, date_formats=['%d %B %Y'])
    time.append(date)

clean_text = [text.replace ("\n", "") for text in news]
clean_text = [" ".join(text.split()) for text in news] 
    
The_Press = pd.DataFrame()
The_Press["filename"] = names
The_Press["text_original"] = news
The_Press["date"] = time
The_Press["year"] = The_Press["date"].dt.year
The_Press["month"] = The_Press["date"].dt.month
The_Press["outlet"] = "The_Press"
The_Press["country"] = "New Zealand"
The_Press["text_clean"] = clean_text
The_Press["South"] = 0

dfs = [Sunday_Times, Globe_Mail, Hindu, The_Nation, NZ_Herald, Star, Times, WPost, Times_IN, TO_Star,
      Bangkok_PO, Sydney_MO, The_Australian, The_Guardian, NY_Times, The_Press]

df = pd.concat(dfs)

df.to_csv("data_new.csv")

len(df)

df.head()
