==> 3_fewshot_classification.ipynb <==
from setfit import SetFitModel, sample_dataset
import pandas as pd
from datasets import load_dataset
import numpy as np
import torch

### Data preperation and model training

dataset = pd.read_csv("scratch/3_SML_benchmark.csv")

dataset.tail()

df = dataset[["text_clean", "RA"]]
df.rename(columns={'RA': 'label','text_clean':'text'}, inplace=True)

#train = df.sample(frac = 0.7)
train = df.groupby('label').sample(n=150)
df2 = df.drop(train.index)
validation = df2.sample(frac = 0.5)
test = df2.drop(validation.index)

train.to_csv("scratch/train.csv")
validation.to_csv("scratch/validation.csv")
test.to_csv("scratch/test.csv")

print(len(train),len(validation),len(test))

dataset = load_dataset("csv", data_files={'train':'scratch/train.csv','validation':'scratch/validation.csv','test': 'scratch/test.csv'})

dataset

#train_dataset = sample_dataset(dataset["train"], label_column="label", num_samples=50)
#train_dataset = dataset["train"]
eval_dataset = dataset["validation"]
test_dataset = dataset["test"]

from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score

# Define custom function for computing metrics
def compute_metrics(y_pred, y_test):
    accuracy = accuracy_score(y_test, y_pred)
    precision = precision_score(y_test, y_pred, average='macro')
    recall = recall_score(y_test, y_pred, average='macro')
    f1 = f1_score(y_test, y_pred, average='macro')
    return {
        'accuracy': accuracy,
        'precision': precision,
        'recall': recall,
        'f1': f1
    }

torch.cuda.empty_cache()

from setfit import TrainingArguments

args = TrainingArguments(
    batch_size=16,
    num_epochs=1,
    max_length=512,
)

from setfit import Trainer

for i in [5, 10, 25, 50, 75, 100, 125, 150]:
    print(i)
    model = SetFitModel.from_pretrained("BAAI/bge-small-en-v1.5") #original retrained model need to be loaded for each round
    train_dataset = sample_dataset(dataset["train"], label_column="label", num_samples=i)
    trainer = Trainer(
        model=model,
        args=args,
        train_dataset=train_dataset,
        metric=compute_metrics,
    )
    trainer.train()
    metrics = trainer.evaluate(eval_dataset)
    print(metrics)

#recheck 100 sample and 150 sample again to see if the results above is subject to chance
from setfit import Trainer

model = SetFitModel.from_pretrained("BAAI/bge-small-en-v1.5")
train_dataset = sample_dataset(dataset["train"], label_column="label", num_samples=100)

trainer = Trainer(
    model=model,
    args=args,
    train_dataset=train_dataset,
    metric=compute_metrics,
)
trainer.train()
metrics = trainer.evaluate(eval_dataset)
print(metrics)

from setfit import Trainer

model = SetFitModel.from_pretrained("BAAI/bge-small-en-v1.5")
train_dataset = sample_dataset(dataset["train"], label_column="label", num_samples=150)

trainer = Trainer(
    model=model,
    args=args,
    train_dataset=train_dataset,
    metric=compute_metrics,
)
trainer.train()
metrics = trainer.evaluate(eval_dataset)
print(metrics)

**As N=100 sample works quite well and there's no huge improvements with introducing more samples, use N=100 sample for fine-tuning**

### Hyperparameter Optimization

from setfit import SetFitModel
from typing import Dict, Any

def model_init(params: Dict[str, Any]) -> SetFitModel:
    params = params or {}
    max_iter = params.get("max_iter", 100)
    solver = params.get("solver", "liblinear")
    params = {
        "head_params": {
            "max_iter": max_iter,
            "solver": solver,
        }
    }
    return SetFitModel.from_pretrained("BAAI/bge-small-en-v1.5", **params)

from optuna import Trial
from typing import Dict, Union

def hp_space(trial: Trial) -> Dict[str, Union[float, int, str]]:
    return {
        "body_learning_rate": trial.suggest_float("body_learning_rate", 1e-5, 1e-3, log=True),
        "num_epochs": trial.suggest_int("num_epochs", 1, 4),
        "batch_size": trial.suggest_categorical("batch_size", [8, 16, 32, 64]),
        "seed": trial.suggest_int("seed", 1, 42),
        "max_iter": trial.suggest_int("max_iter", 50, 300),
        "solver": trial.suggest_categorical("solver", ["newton-cg", "lbfgs", "liblinear"]),
    }

#train_dataset = sample_dataset(dataset["train"], label_column="label", num_samples=100)

trainer = Trainer(
    train_dataset=train_dataset,
    eval_dataset=eval_dataset,
    model_init=model_init,
    metric=compute_metrics,
)
best_run = trainer.hyperparameter_search(direction="maximize", hp_space=hp_space, n_trials=10)

print(best_run)

trainer.apply_hyperparameters(best_run.hyperparameters, final_model=True)
trainer.train()

trainer.evaluate(eval_dataset)

trainer.evaluate(test_dataset)

### Making predictions

df = pd.read_csv("scratch/2_data_preperation_data_dedup.csv")
df.drop(labels = ["Unnamed: 0"], axis = 1, inplace = True)

len(df)

df.head()

news_pred = df["text_clean"]

len(news_pred)

preds = model.predict(news_pred)

df["keep"] = preds

df["keep"].value_counts()

pre = df.loc[df["keep"]==1]

len(pre)

pre.to_csv("scratch/3_fewshot_classification_data_final.csv")
