==> 4_Inductive_label _identification.ipynb <==
import pandas as pd
from nltk.tokenize import word_tokenize
from nltk import bigrams
import re
from collections import Counter

df = pd.read_csv('3_fewshot_classification_data_final.csv')
df.drop(labels = ['Unnamed: 0'], axis = 1, inplace = True)
pd.set_option('display.max_columns', None)

df.head()

texts = df["text_clean"].to_list()
texts =[x.lower() for x in texts]
pattern = re.compile(r"climat[e]*|heat[ing]*|warm[ing]*|greenhous[e]", re.IGNORECASE)

all_bigrams = []

for article in texts:
    matches = re.findall(pattern, article)
    words = word_tokenize(article)
    target_bigrams = list(bigrams(words))
    target_bigrams = [(word1, word2) for word1, word2 in target_bigrams if word1 in matches or word2 in matches]
    all_bigrams.extend(target_bigrams)

element_counts = Counter(all_bigrams)
filtered_list = [element for element, count in element_counts.items() if count >= 100]

len(filtered_list)

filtered_list

def extract_sentence(text, query):
    #import re
    #text = str(text).lower()
    sentences = ""
    #query = str(query).lower()
    
    for sentence in text.split('.'):
        newsen = re.sub(r'[^\w\s.]', ' ', sentence)
        if query in newsen:
            sentences = sentences+sentence
    return sentences

names = ["changing climate", 'climate risk', 'climate problem', 'severe climate', 'climate extremes', 
         'extreme climate', 'dangerous climate', 'climate challenge', 'climate risks', 'climate wars',
        'climate issue', 'climate issues','climate concerns', 'climate changes', 'catastrophic climate', 
         'climate hazards', 'climate threats', 'climate catastrophes', 'warmer climate', 'climate disasters', 
         'climate damage', 'catastrophic warming', 'planetary warming', 'climate warms']

sentences_dict = {}

for name in names:
    sentences_dict[name] = []
    for article in texts:
        extracted_sen = extract_sentence(article, name)
        if len(extracted_sen) != 0:
            sentences_dict[name].append(extracted_sen)

import pickle

with open('4_Inductive_label _identification_sentences.pkl', 'wb') as fp:
    pickle.dump(sentences_dict, fp)

sentences_dict["climate warms"]

len(sentences_dict["climate issue"])
